backend/requirements.txt
aiohttp
aiolimiter
beautifulsoup4
feedparser
html5lib
lxml
pydantic
pyyaml
python-dotenv
readability-lxml
tqdm
openai
sentence-transformers
scikit-learn
sentry-sdk

fastapi
uvicorn[standard]

# Testing
pytest
pytest-asyncio
pytest-cov